Move `formatDate()` to a trait
This function was used separately in 6 distinct subdecisions. This
can only to issues where the function definition differs per
subdecision. As such, it has been moved to a trait to provide one
definition for all subdecisions.
