+++ module/Database/src/Model/SubDecision/Board/Release.php
@@ -5,14 +5,12 @@ declare(strict_types=1);
 namespace Database\Model\SubDecision\Board;
 
 use Database\Model\SubDecision;
+use Database\Model\Trait\FormattableDateTrait;
 use DateTime;
 use Doctrine\ORM\Mapping\Column;
 use Doctrine\ORM\Mapping\Entity;
 use Doctrine\ORM\Mapping\JoinColumn;
 use Doctrine\ORM\Mapping\OneToOne;
-use IntlDateFormatter;
-
-use function date_default_timezone_get;
 
 /**
  * Release from board duties.
@@ -23,6 +21,8 @@ use function date_default_timezone_get;
 #[Entity]
 class Release extends SubDecision
 {
+    use FormattableDateTrait;
+
     /**
      * Reference to the installation of a member.
      */
@@ -88,29 +88,6 @@ class Release extends SubDecision
     public function setDate(DateTime $date): void
     {
         $this->date = $date;
-    }
-
-    /**
-     * Format the date.
-     *
-     * returns the localized version of $date->format('d F Y')
-     *
-     * @return string Formatted date
-     */
-    protected function formatDate(
-        DateTime $date,
-        string $locale = 'nl_NL',
-    ): string {
-        $formatter = new IntlDateFormatter(
-            $locale,
-            IntlDateFormatter::NONE,
-            IntlDateFormatter::NONE,
-            date_default_timezone_get(),
-            null,
-            'd MMMM y',
-        );
-
-        return $formatter->format($date);
     }
 
     protected function getTemplate(): string
